backend/requirements.txt
fastapi
uvicorn[standard]
langchain
openai
chromadb
pytesseract
pypdf
sqlalchemy
python-dotenv
tqdm
pillow
pdfminer.six
pydantic-settings
langchain-community
python-multipart
tiktoken
starlette
google-generativeai
langchain_google_genai